requirements.txt
altgraph==0.17.2
Cython==0.29.32
future==0.18.2
pefile==2022.5.30
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyjnius==1.4.2
PySide6==6.3.2
PySide6-Addons==6.3.2
PySide6-Essentials==6.3.2
pywin32-ctypes==0.2.0
shiboken6==6.3.2
six==1.16.0
